% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataModel.R
\name{DataModel}
\alias{DataModel}
\title{Fmmr6 DataModel Object}
\value{
Returns R6 object of class DataModel.
}
\description{
A concrete R6 object to generate data based on the formula
to use in the fmmr6.
}
\author{
Dongjie Wu
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formula}}{(\code{formula(1)})\cr
The formula used in the model.}

\item{\code{data_format}}{(\code{character(1)})\cr
The data format: default is "data.frame".}

\item{\code{data}}{(\code{data.frame()|data.table()|table()})\cr
Data used in the model.}

\item{\code{X}}{(\code{matrix()})\cr
The matrix of the independent variables.}

\item{\code{Y}}{(\code{matrix()})\cr
The matrix of the dependent variables.}

\item{\code{Y_coln}}{(\code{character()})\cr
The column value labels.}

\item{\code{family}}{(\code{character(1)})\cr
The distribution family of the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataModel$new()}}
\item \href{#method-clone}{\code{DataModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize and generate the \code{DataModel} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataModel$new(data, formula, family, mn_base = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(\code{data.frame()|data.table()|table()})\cr
Data used in the model.}

\item{\code{formula}}{(\code{formula(1)})\cr
The formula used in the model.}

\item{\code{family}}{(\code{character(1)})\cr
The distribution family of the model.}

\item{\code{mn_base}}{(\code{integer(1)}) \cr
Determine which column of the multinomial variable is set to be the base group.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
