% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimMethod.R
\name{EstimMethod}
\alias{EstimMethod}
\title{An Abstract Class for Estimation Methods}
\value{
No return.
}
\description{
An abstract class for estimation methods which cannot be initialized.
}
\author{
Dongjie Wu
}
\section{Super class}{
\code{\link[fmmr6:AbstractMethod]{fmmr6::AbstractMethod}} -> \code{EstimMethod}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data_model}}{(\code{DataModel()}) \cr
The DataModel Object that stores the data using in the fmmr6.}

\item{\code{constraint}}{(\code{matrix()}) \cr
The constraint matrix.}

\item{\code{latent}}{(\code{integer(1)}) \cr
The number of latent classes}

\item{\code{optim_method}}{(\code{character(1)}) \cr
The optimization method to use to fit the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EstimMethod$new()}}
\item \href{#method-partial}{\code{EstimMethod$partial()}}
\item \href{#method-gen_start}{\code{EstimMethod$gen_start()}}
\item \href{#method-gen_constraint}{\code{EstimMethod$gen_constraint()}}
\item \href{#method-clone}{\code{EstimMethod$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="AbstractMethod" data-id="fit">}\href{../../fmmr6/html/AbstractMethod.html#method-fit}{\code{fmmr6::AbstractMethod$fit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link{R6} \link[R6:R6Class]{R6::R6Class} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EstimMethod$new(
  latent,
  data_model,
  start = NULL,
  optim_method = "base",
  use_llc = T,
  constraint = matrix(1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{latent}}{(\code{numeric(1)}) \cr
The number of latent classes.}

\item{\code{data_model}}{(\code{DataModel()}) \cr
The DataModel object contains data used in the fmmr6.}

\item{\code{start}}{(\code{matrix()}) \cr
The matrix of start values for the EM algorithm.}

\item{\code{optim_method}}{(\code{character(1)}) \cr
The optimization method to use to fit the model.
The default is \code{base}.}

\item{\code{use_llc}}{(\code{boolean(1)}) \cr
Whether to use the complete log-likelihood or the normal log-likelihood.
The default is \code{TRUE}.}

\item{\code{constraint}}{(\code{matrix()}) \cr
The matrix of constraint values for the EM algorithm}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return a R6 object of class em.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-partial"></a>}}
\if{latex}{\out{\hypertarget{method-partial}{}}}
\subsection{Method \code{partial()}}{
Generate the partial function by filling in some parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EstimMethod$partial(f, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{(\verb{function()}) \cr
The function to be filled in.}

\item{\code{...}}{(\code{list()}) \cr
The list of the parameters to filled in to the functions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_start"></a>}}
\if{latex}{\out{\hypertarget{method-gen_start}{}}}
\subsection{Method \code{gen_start()}}{
Generate the start values of parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EstimMethod$gen_start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_constraint"></a>}}
\if{latex}{\out{\hypertarget{method-gen_constraint}{}}}
\subsection{Method \code{gen_constraint()}}{
Generate the constraint of parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EstimMethod$gen_constraint()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EstimMethod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
