% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Family.R
\name{Family}
\alias{Family}
\title{An Abstract Class for Distribution Families}
\value{
No return.
}
\description{
An abstract class for Distribution Families
which cannot be initialized.
}
\author{
Dongjie Wu
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Family$new()}}
\item \href{#method-gen_ll}{\code{Family$gen_ll()}}
\item \href{#method-gen_gr}{\code{Family$gen_gr()}}
\item \href{#method-clone}{\code{Family$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Family$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The function will return an error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_ll"></a>}}
\if{latex}{\out{\hypertarget{method-gen_ll}{}}}
\subsection{Method \code{gen_ll()}}{
Generate a log-likelihood function for the
distribution family.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Family$gen_ll()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_gr"></a>}}
\if{latex}{\out{\hypertarget{method-gen_gr}{}}}
\subsection{Method \code{gen_gr()}}{
Generate a gradient function for the
distribution family.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Family$gen_gr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Family$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
