% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamilyLogit.R
\name{FamilyLogit}
\alias{FamilyLogit}
\title{Family Object for a Logistic Distribution}
\value{
A FamilyLogit Object
}
\description{
An family object for a
logistic distribution.
}
\author{
Dongjie Wu
}
\section{Super class}{
\code{\link[fmmr6:Family]{fmmr6::Family}} -> \code{FamilyLogit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FamilyLogit$new()}}
\item \href{#method-gen_density}{\code{FamilyLogit$gen_density()}}
\item \href{#method-gen_start}{\code{FamilyLogit$gen_start()}}
\item \href{#method-gen_constraint}{\code{FamilyLogit$gen_constraint()}}
\item \href{#method-clone}{\code{FamilyLogit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyLogit$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_density"></a>}}
\if{latex}{\out{\hypertarget{method-gen_density}{}}}
\subsection{Method \code{gen_density()}}{
Generate the density function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyLogit$gen_density()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{(\code{matrix()}) \cr
A matrix with 1 column contains the dependent variable.}

\item{\code{X}}{(\code{matrix()}) \cr
A matrix with m column contains m independent variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
this function returns the density function of the logistic regression.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_start"></a>}}
\if{latex}{\out{\hypertarget{method-gen_start}{}}}
\subsection{Method \code{gen_start()}}{
Generate the start value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyLogit$gen_start(Y, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{(\code{matrix()}) \cr
A matrix with 1 column contains the dependent variable.}

\item{\code{X}}{(\code{matrix()}) \cr
A matrix with m column contains m independent variables.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_constraint"></a>}}
\if{latex}{\out{\hypertarget{method-gen_constraint}{}}}
\subsection{Method \code{gen_constraint()}}{
Generate the constraint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyLogit$gen_constraint(Y, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{(\code{matrix()}) \cr
A matrix with 1 column contains the dependent variable.}

\item{\code{X}}{(\code{matrix()}) \cr
A matrix with m column contains m independent variables.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyLogit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
