% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamilyMNormal.R
\name{FamilyMNormal}
\alias{FamilyMNormal}
\title{Family Object for Mixture of Normal Distributions}
\value{
A FamilyMNormal Object
}
\description{
An family object for mixture of
normal distributions.
}
\author{
Dongjie Wu
}
\section{Super class}{
\code{\link[fmmr6:Family]{fmmr6::Family}} -> \code{FamilyMNormal}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FamilyMNormal$new()}}
\item \href{#method-gen_ll}{\code{FamilyMNormal$gen_ll()}}
\item \href{#method-clone}{\code{FamilyMNormal$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="Family" data-id="gen_gr">}\href{../../fmmr6/html/Family.html#method-gen_gr}{\code{fmmr6::Family$gen_gr()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyMNormal$new(Y, X, latent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_ll"></a>}}
\if{latex}{\out{\hypertarget{method-gen_ll}{}}}
\subsection{Method \code{gen_ll()}}{
Generate the likelihood function for a mixture of
normal distributions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyMNormal$gen_ll()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a function \code{ll} with two parameters,
\code{theta} is a vector with parameters whose number
is number of variables in X + 2 times
number of latent classes.
\code{d} is the indicator matrix whose size is:
rows of X by number of latent classes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyMNormal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
