% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamilyMultiNomial.R
\name{FamilyMultiNomial}
\alias{FamilyMultiNomial}
\title{Family Object for a MultiNomial Distribution}
\value{
A FamilyMultiNomial Object
}
\description{
An family object for a
MultiNomial distribution.
}
\author{
Dongjie Wu
}
\section{Super class}{
\code{\link[fmmr6:Family]{fmmr6::Family}} -> \code{FamilyMultiNomial}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FamilyMultiNomial$new()}}
\item \href{#method-gen_density}{\code{FamilyMultiNomial$gen_density()}}
\item \href{#method-gen_start}{\code{FamilyMultiNomial$gen_start()}}
\item \href{#method-gen_constraint}{\code{FamilyMultiNomial$gen_constraint()}}
\item \href{#method-clone}{\code{FamilyMultiNomial$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyMultiNomial$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_density"></a>}}
\if{latex}{\out{\hypertarget{method-gen_density}{}}}
\subsection{Method \code{gen_density()}}{
Generate the density function: \cr
Assume that there are \eqn{N} observations,
\eqn{M} variables and \eqn{P} choices.
\eqn{X} is \eqn{N \times M},
\eqn{Y} is \eqn{N \times P} and
\eqn{\theta} is \eqn{M \times P}
Then the density function is:
\deqn{R_k = \sum_{i \in N}(y_{ik} \times 
\sum_{j \in M}(x_{ij} \times \theta_{jk})) - 
\log(1+\sum_{i \in N}(\exp{\sum_{j \in M}(x_{ij} \times \theta_{jk})})}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyMultiNomial$gen_density()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_start"></a>}}
\if{latex}{\out{\hypertarget{method-gen_start}{}}}
\subsection{Method \code{gen_start()}}{
Generate the start value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyMultiNomial$gen_start(Y, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{\code{matrix(1)} \cr
A matrix with 1 column contains the dependent variable.}

\item{\code{X}}{\code{matrix(1)} \cr
A matrix with m column contains m independent variables.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_constraint"></a>}}
\if{latex}{\out{\hypertarget{method-gen_constraint}{}}}
\subsection{Method \code{gen_constraint()}}{
Generate the constraint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyMultiNomial$gen_constraint(Y, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{\code{matrix(1)} \cr
A matrix with 1 column contains the dependent variable.}

\item{\code{X}}{\code{matrix(1)} \cr
A matrix with m column contains m independent variables.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyMultiNomial$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
