% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FamilyNormal.R
\name{FamilyNormal}
\alias{FamilyNormal}
\title{Family Object for a Normal Distribution}
\value{
A FamilyNormal Object
}
\description{
An family object for a
normal distribution.
}
\author{
Dongjie Wu
}
\section{Super class}{
\code{\link[fmmr6:Family]{fmmr6::Family}} -> \code{FamilyNormal}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FamilyNormal$new()}}
\item \href{#method-gen_ll}{\code{FamilyNormal$gen_ll()}}
\item \href{#method-gen_gr}{\code{FamilyNormal$gen_gr()}}
\item \href{#method-clone}{\code{FamilyNormal$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyNormal$new(Y, X, latent = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_ll"></a>}}
\if{latex}{\out{\hypertarget{method-gen_ll}{}}}
\subsection{Method \code{gen_ll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyNormal$gen_ll()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_gr"></a>}}
\if{latex}{\out{\hypertarget{method-gen_gr}{}}}
\subsection{Method \code{gen_gr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyNormal$gen_gr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FamilyNormal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
