% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmglm.R
\name{fmglm}
\alias{fmglm}
\title{Finite Mixture of Generalized Linear Models (fmglm) Object}
\value{
Returns R6 object of class fmglm.
}
\description{
A finite mixture of generalized linear model (GLM)
}
\author{
Dongjie Wu
}
\section{Super class}{
\code{\link[fmmr6:fmmr6]{fmmr6::fmmr6}} -> \code{fmglm}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fmglm$new()}}
\item \href{#method-fit}{\code{fmglm$fit()}}
\item \href{#method-summarize}{\code{fmglm$summarize()}}
\item \href{#method-clone}{\code{fmglm$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="fmmr6" data-id="predict">}\href{../../fmmr6/html/fmmr6.html#method-predict}{\code{fmmr6::fmmr6$predict()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link{R6} \link[R6:R6Class]{R6::R6Class} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmglm$new(
  formula,
  data,
  family = "gaussian",
  latent = 2,
  method = "em",
  start = NULL,
  glm_fit = F,
  use_llc = T,
  mn_base = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{\code{formula(1)} \cr
The formula/expression of the model to fit in the fmglm}

\item{\code{data}}{\code{data(1)} \cr
The Data used in the fmglm}

\item{\code{family}}{\code{character(1)} \cr
The distribution family which can be either a string like "gaussian"
or a vector like \code{c("gaussian", "poisson")}}

\item{\code{latent}}{\code{numeric(1)} \cr
The number of latent classes}

\item{\code{method}}{\code{character(1)} \cr
The estimation method to fit the fmglm}

\item{\code{start}}{\code{matrix(1)} \cr
The starting values for the fmglm}

\item{\code{glm_fit}}{\code{Boolean(1)} \cr
Whether use the \code{glm.fit()}, \code{ols.wfit()} or \code{nnet()} to fit the model.
The default is \code{FALSE}.}

\item{\code{use_llc}}{\code{Boolean(1)} \cr
Whether to use the complete log-likelihood to fit the model.
The default is \code{TRUE}.}

\item{\code{mn_base}}{\code{integer(1)} \cr
Determine which column of the multinomial variable is set to be the base group.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return a R6 object of class fmglm
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the fmglm model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmglm$fit(algo = "em", max_iter = 500, start = "kmeans", rep = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{algo}}{\code{character(1)} \cr
The algorithm used in fitting the fmglm model.
The default algorithm is \code{em} standing for the normal EM algorithm.
One can choose from \code{c("em", "cem", "sem")}.
\code{cem} is the classification EM algorithm.
\code{sem} is the stochastic EM algorithm.}

\item{\code{max_iter}}{\code{numeric(1)} \cr
Specify the maximum number of iterations for the E-step-M-step loop.
The default number is 500.}

\item{\code{start}}{\code{character(1)} \cr
Specify the starting method of the EM algorithm.
Can either start from \code{kmeans} or \code{random}.
\code{kmeans} use the K-mean methods to put samples into latent classes.
\code{random} randomly assigns samples into latent classes.
The default method is \code{kmeans}.}

\item{\code{rep}}{\code{numeric(1)} \cr
Specify the number of reps EM-algorithm runs.
This parameter is designed for preventing the local maximum.
Each rep, the EM_algorithm generates a start.
It is only useful when \code{start} is \code{random}.
After all reps, the algorithm will pick the rep with maximum log likelihood.
The default value is 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
Generate a summary for the result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmglm$summarize(digits = 3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{\code{integer(1)} \cr
Determine how many digits presented in the output.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmglm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
