% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmmr6.R
\name{fmmr6}
\alias{fmmr6}
\title{The Abstract Class of Finite Mixture Models on R6 (fmmr6)}
\value{
No return
}
\description{
An abstract R6Class for the Finite Mixture Models.
}
\section{The R6 Structure of the \code{fmmr6} Package}{

The \code{fmmr6} package adopts an Object-Oriented (OO) design with the help of \CRANpkg{R6}.
It contains three main blocks of constructing and fitting a finite mixture model (FMM):
\itemize{
\item Finite Mixture Modelling:
Finite Mixture Modelling can the way we construct our finite mixture models.
These modelling techniques are organized under \code{fmmr6} class.
In the current version, we have only developed the mixture of generalized linear models,
which is the \code{fmglm} class.
We are going to add more modelling techniques such as the mixture of
generalized linear models using a panel data structure \code{xtfmglm}.
\item Methods to fit a FMM models:
The fitting methods such as EM-algorithm (\code{em}) and Maximum Likelihood Estimation (\code{mle})
are organized under the \code{AbstractMethod} class. In this way, it is convenient to pickup
a different method to fit the FMM models.
\item Families of distributions:
Different families of distributions are organized under the \code{Family} class.
Currently, the package only supports Gaussian distributions (\code{FamilyNormal}),
Poisson distributions (\code{FamilyPoisson}),
Logit Models with Gaussian distributions (\code{FamilyLogit}),
and Multinomial Logit Models with Gaussian distributions (\code{FamilyMultiNomial}).
These distributions will be sent to a \code{Mixer} class to generate a mixture of
distributions based on the setting of the finite mixture model.
}
}

\author{
Dongjie Wu
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fmmr6$new()}}
\item \href{#method-fit}{\code{fmmr6$fit()}}
\item \href{#method-summarize}{\code{fmmr6$summarize()}}
\item \href{#method-predict}{\code{fmmr6$predict()}}
\item \href{#method-clone}{\code{fmmr6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link{R6} \link[R6:R6Class]{R6::R6Class} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmmr6$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the fmmr6 model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmmr6$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
Generate a summary for the result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmmr6$summarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict the outcome given the fitted model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmmr6$predict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmmr6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
