% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtab.R
\name{fmtab}
\alias{fmtab}
\title{Finite Mixture of Contingency Table Analysis (fmtab) Object}
\value{
Returns R6 object of class fmtab.
}
\description{
A finite mixture of models related to contingency table analysis
}
\author{
Dongjie Wu
}
\section{Super class}{
\code{\link[fmmr6:fmmr6]{fmmr6::fmmr6}} -> \code{fmtab}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data_model}}{(\code{DataModel()}) \cr
The DataModel Object that stores the data using in the fmmr6.}

\item{\code{family}}{(\code{character(1)|character()}) \cr
The distribution family which can be either a string like "gaussian"
or a vector like \code{c("gaussian", "gaussian")}.}

\item{\code{latent}}{(\code{integer(1)}) \cr
The number of latent classes.}

\item{\code{method}}{(\code{character(1)}) \cr
The estimation method to fit the fmglm.}

\item{\code{start}}{(\code{matrix()}) \cr
The starting values for the fmglm.}

\item{\code{constraint}}{(\code{matrix()}) \cr
The constraint matrix.}

\item{\code{concomitant}}{(\code{formula(1)})\cr
The formula to model the concomitant model.
The default value is NULL.}

\item{\code{optim_method}}{(\code{character(1)}) \cr
The optimization method to use to fit the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fmtab$new()}}
\item \href{#method-fit}{\code{fmtab$fit()}}
\item \href{#method-summarize}{\code{fmtab$summarize()}}
\item \href{#method-clone}{\code{fmtab$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="fmmr6" data-id="predict">}\href{../../fmmr6/html/fmmr6.html#method-predict}{\code{fmmr6::fmmr6$predict()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link{R6} \link[R6:R6Class]{R6::R6Class} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmtab$new(
  formula,
  data,
  data_str = "longc",
  data_var = NULL,
  model = "loglinear",
  latent = 2,
  method = "em",
  start = NULL,
  optim_method = "base",
  concomitant = NULL,
  constraint = matrix(1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{(\code{formula(1)}) \cr
The formula/expression of the model to fit in the fmglm.}

\item{\code{data}}{(\code{data.frame()}) \cr
The Data used in the fmglm.}

\item{\code{data_str}}{(\code{character(1)})\cr
Desired data structure.
Available data structure are listed as follows:

\code{default}: Default data structure, the DataModel object will identify
the dependent variable \code{Y} and the independent variable \code{X} based on
the provided formula.

\code{freq_tab}: the frequency table. The data must contain a \code{freq} variable.

\code{long}: Data in the long format. Normally used for a longitudinal/panel data.
The data must contain a variable to identify the changes within observation,
e.g. the observations made at different time \code{t}. One should also define a variable
for identifying each observation, e.g. \code{obs}.

\code{wide}: Data in the wide format. This data format collapses a longitudinal data
into one row for one observation.

\code{longc}: Data in the long format used in a choice model. Variable \code{alt} represents
the alternative for each individual. Variable \code{chosen} shows which \code{alt} is chosen.

\code{widec} Data in the wide format used in a choice model.}

\item{\code{latent}}{(\code{integer(1)}) \cr
The number of latent classes.}

\item{\code{method}}{(\code{character(1)}) \cr
The estimation method to fit the fmglm.}

\item{\code{start}}{(\code{matrix()}) \cr
The starting values for the fmglm.}

\item{\code{optim_method}}{(\code{character(1)}) \cr
The optimization method to use to fit the model.
The default is \code{base}.}

\item{\code{concomitant}}{(\code{formula(1)}) \cr
The formula for the concomitant model. E.g. \code{~ z1 + z2 + z3}.}

\item{\code{constraint}}{(\code{matrix()}) \cr
The constraint matrix.}

\item{\code{use_llc}}{(\code{boolean(1)}) \cr
Whether to use the complete log-likelihood or the normal log-likelihood.
The default is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return a R6 object of class fmglm
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the fmglm model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmtab$fit(algo = "em", max_iter = 500, start = "random", rep = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{algo}}{(\code{character(1)}) \cr
The algorithm used in fitting the fmglm model.
The default algorithm is \code{em} standing for the normal EM algorithm.
One can choose from \code{c("em", "cem", "sem")}.
\code{cem} is the classification EM algorithm.
\code{sem} is the stochastic EM algorithm.}

\item{\code{max_iter}}{(\code{integer(1)}) \cr
Specify the maximum number of iterations for the E-step-M-step loop.
The default number is 500.}

\item{\code{start}}{(\code{character(1)}) \cr
Specify the starting method of the EM algorithm.
Can either start from \code{kmeans} or \code{random}.
\code{kmeans} use the K-mean methods to put samples into latent classes.
\code{random} randomly assigns samples into latent classes.
The default method is \code{kmeans}.}

\item{\code{rep}}{(\code{integer(1)}) \cr
Specify the number of reps EM-algorithm runs.
This parameter is designed for preventing the local maximum.
Each rep, the EM_algorithm generates a start.
It is only useful when \code{start} is \code{random}.
After all reps, the algorithm will pick the rep with maximum log likelihood.
The default value is 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
Generate a summary for the result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmtab$summarize(digits = 3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{(\code{integer(1)}) \cr
Determine how many digits presented in the output.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmtab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
