% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfmm.R
\name{glfmm}
\alias{glfmm}
\title{Generalized Linear Finite Mixture Models (GLFMMs) Object}
\value{
Returns R6 object of class glfmm.
}
\description{
A finite mixture of generalized linear model (GLM)
}
\author{
Dongjie Wu
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{glfmm$new()}}
\item \href{#method-summary}{\code{glfmm$summary()}}
\item \href{#method-fit}{\code{glfmm$fit()}}
\item \href{#method-predict}{\code{glfmm$predict()}}
\item \href{#method-clone}{\code{glfmm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link{R6} \link[R6:R6Class]{R6::R6Class} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{glfmm$new(
  formula,
  data,
  link = "identity",
  latent = 1,
  method = "mle",
  start = NULL,
  family = "gaussian"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{\code{formula(1)} \cr
The formula of the model to fit in the glfmm}

\item{\code{data}}{\code{data(1)} \cr
The Data used in the glfmm}

\item{\code{link}}{\code{character(1)} \cr
The link function in the glfmm}

\item{\code{latent}}{\code{numeric(1)} \cr
The number of latent classes}

\item{\code{method}}{\code{character(1)} \cr
The estimation method to fit the glfmm}

\item{\code{start}}{\code{matrix(1)} \cr
The starting values for the glfmm}

\item{\code{family}}{\code{character(1)} \cr
The distribution family}

\item{\code{summary}}{\verb{function(1)} \cr
Print the detailed model-fitting results}

\item{\code{fit}}{\verb{function(1)} \cr
Fit the model}

\item{\code{predict}}{\verb{function(1)} \cr
Use the model to predict the dependent variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{glfmm$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{glfmm$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{glfmm$predict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{glfmm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
