% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss94}
\alias{gss94}
\title{A dataset containing siblings and occupation data.}
\format{
A data frame with 2236 rows and 12 variables:
}
\usage{
gss94
}
\description{
A dataset containing siblings and occupation data.
}
\keyword{datasets}
