% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss94_educ}
\alias{gss94_educ}
\title{A dataset containing siblings and education data.}
\format{
A data frame with 3696 rows and 13 variables:
}
\usage{
gss94_educ
}
\description{
A dataset containing siblings and education data.
}
\keyword{datasets}
