% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimBase.R
\name{OptimBase}
\alias{OptimBase}
\title{A Class for the GLM Optimization Method.}
\value{
No return.
}
\description{
A class for the optimization methods to use ´optim()´ in the R base.
}
\author{
Dongjie Wu
}
\section{Super classes}{
\code{\link[fmmr6:AbstractMethod]{fmmr6::AbstractMethod}} -> \code{\link[fmmr6:OptimMethod]{fmmr6::OptimMethod}} -> \code{OptimBase}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fit}{\code{OptimBase$fit()}}
\item \href{#method-clone}{\code{OptimBase$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="OptimMethod" data-id="initialize">}\href{../../fmmr6/html/OptimMethod.html#method-initialize}{\code{fmmr6::OptimMethod$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Run the optimization for the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimBase$fit(
  data_model,
  theta,
  ll,
  gr,
  hidden,
  pi_vector,
  npar,
  latent,
  family
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_model}}{(\code{DataModel()}) \cr
The DataModel object contains data used in the fmmr6.}

\item{\code{theta}}{(\code{numeric()}) \cr
The coefficients to estimates.}

\item{\code{ll}}{(\verb{function()}) \cr
The loglikelihood function.}

\item{\code{gr}}{(\verb{function()}) \cr
The gradient function.}

\item{\code{hidden}}{(\code{matrix()}) \cr
The matrix of the posterior probability.}

\item{\code{pi_vector}}{(\code{numeric()}) \cr
A vector of the prior probability \code{pi}.}

\item{\code{npar}}{(\code{integer()}) \cr
Number of the parameters.}

\item{\code{latent}}{(\code{integer(1)}) \cr
The number of latent classes}

\item{\code{family}}{(\code{character(1)|character()}) \cr
The distribution family which can be either a string like "gaussian"
or a vector like \code{c("gaussian", "gaussian")}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return the optimization result with the estimates,
the Loglikelihood value and the information criteria.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
