% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimMethod.R
\name{OptimMethod}
\alias{OptimMethod}
\title{A Class for Optimization Methods}
\value{
No return.
}
\description{
A class for optimization methods which cannot be initialized.
}
\author{
Dongjie Wu
}
\section{Super class}{
\code{\link[fmmr6:AbstractMethod]{fmmr6::AbstractMethod}} -> \code{OptimMethod}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimMethod$new()}}
\item \href{#method-clone}{\code{OptimMethod$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="AbstractMethod" data-id="fit">}\href{../../fmmr6/html/AbstractMethod.html#method-fit}{\code{fmmr6::AbstractMethod$fit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
The initialization of the optimization methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimMethod$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimMethod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
