% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{EM Algorithm (EM) Object}
\value{
Returns R6 object of class em.
}
\description{
A concrete R6 object for EM Algorithm.
}
\author{
Dongjie Wu
}
\section{Super class}{
\code{\link[fmmr6:AbstractMethod]{fmmr6::AbstractMethod}} -> \code{em}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{em$new()}}
\item \href{#method-fit}{\code{em$fit()}}
\item \href{#method-estep}{\code{em$estep()}}
\item \href{#method-mstep}{\code{em$mstep()}}
\item \href{#method-clone}{\code{em$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link{R6} \link[R6:R6Class]{R6::R6Class} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{em$new(likelihood_func, Y, X, latent, start = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{likelihood_func}}{\code{formula(1)} \cr
The Likelihood function to be used in the EM algorithm.}

\item{\code{start}}{\code{vector(1)} \cr
The vector of start values for the EM algorithm.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return a R6 object of class em.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{em$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estep"></a>}}
\if{latex}{\out{\hypertarget{method-estep}{}}}
\subsection{Method \code{estep()}}{
Given the data, start values, the number
of latent classes, compute the expected value of
the indicator variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{em$estep(Y, X, latent, start)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mstep"></a>}}
\if{latex}{\out{\hypertarget{method-mstep}{}}}
\subsection{Method \code{mstep()}}{
Given the data, start values, the number
of latent classes, the indicator variable, the likelihood
function, solve the maximum value of the likelihood
function and update the parameter vectors theta.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{em$mstep(Y, X, latent, d, start, likelihood_func)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{em$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
