% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Family.R
\name{Family}
\alias{Family}
\title{An Abstract Class for Distribution Families}
\value{
No return.
@description
The initialize function for the Family class:
This class cannot be initialized
}
\description{
An abstract class for Distribution Families
which cannot be initialized.
}
\author{
Dongjie Wu
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Family$new()}}
\item \href{#method-gen_density}{\code{Family$gen_density()}}
\item \href{#method-gen_start}{\code{Family$gen_start()}}
\item \href{#method-gen_constraint}{\code{Family$gen_constraint()}}
\item \href{#method-clone}{\code{Family$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Family$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The function will return an error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_density"></a>}}
\if{latex}{\out{\hypertarget{method-gen_density}{}}}
\subsection{Method \code{gen_density()}}{
Generate the density function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Family$gen_density()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_start"></a>}}
\if{latex}{\out{\hypertarget{method-gen_start}{}}}
\subsection{Method \code{gen_start()}}{
Generate the start value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Family$gen_start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gen_constraint"></a>}}
\if{latex}{\out{\hypertarget{method-gen_constraint}{}}}
\subsection{Method \code{gen_constraint()}}{
Generate the constraint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Family$gen_constraint()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Family$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
