% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmglm.R
\name{fmglm}
\alias{fmglm}
\title{Finite Mixture of Generalized Linear Models (fmglm) Object}
\value{
Returns R6 object of class fmglm.
}
\description{
A finite mixture of generalized linear model (GLM)
}
\author{
Dongjie Wu
}
\section{Super class}{
\code{\link[fmmr6:fmmr6]{fmmr6::fmmr6}} -> \code{fmglm}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fmglm$new()}}
\item \href{#method-fit}{\code{fmglm$fit()}}
\item \href{#method-clone}{\code{fmglm$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="fmmr6" data-id="predict">}\href{../../fmmr6/html/fmmr6.html#method-predict}{\code{fmmr6::fmmr6$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="fmmr6" data-id="summarize">}\href{../../fmmr6/html/fmmr6.html#method-summarize}{\code{fmmr6::fmmr6$summarize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link{R6} \link[R6:R6Class]{R6::R6Class} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmglm$new(
  formula,
  data,
  link = "identity",
  latent = 1,
  method = "mle",
  start = NULL,
  intercept = TRUE,
  family = "gaussian"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{\code{formula(1)} \cr
The formula/expression of the model to fit in the fmglm}

\item{\code{data}}{\code{data(1)} \cr
The Data used in the fmglm}

\item{\code{link}}{\code{character(1)} \cr
The link function in the fmglm}

\item{\code{latent}}{\code{numeric(1)} \cr
The number of latent classes}

\item{\code{method}}{\code{character(1)} \cr
The estimation method to fit the fmglm}

\item{\code{start}}{\code{matrix(1)} \cr
The starting values for the fmglm}

\item{\code{intercept}}{\code{bool(1)} \cr
Whether the model includes an intercept,
the default is TRUE.}

\item{\code{family}}{\code{character(1)} \cr
The distribution family}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return a R6 object of class fmglm
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the fmglm model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmglm$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmglm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
