% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmmr6.R
\name{fmmr6}
\alias{fmmr6}
\title{The Abstract Class of Finite Mixture Models on R6 (fmmr6)}
\value{
No return
}
\description{
An abstract R6Class for the Finite Mixture Models.
}
\author{
Dongjie Wu
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fmmr6$new()}}
\item \href{#method-fit}{\code{fmmr6$fit()}}
\item \href{#method-summarize}{\code{fmmr6$summarize()}}
\item \href{#method-predict}{\code{fmmr6$predict()}}
\item \href{#method-clone}{\code{fmmr6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link{R6} \link[R6:R6Class]{R6::R6Class} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmmr6$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the fmmr6 model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmmr6$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
Generate a summary for the result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmmr6$summarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict the outcome given the fitted model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmmr6$predict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmmr6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
