% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Maximum Likelihood Estimation (MLE) Object}
\value{
Returns R6 object of class mle.
}
\description{
A concrete R6 object for Maximum Likelihood Estimation.
}
\author{
Dongjie Wu
}
\section{Super classes}{
\code{\link[fmmr6:AbstractMethod]{fmmr6::AbstractMethod}} -> \code{\link[fmmr6:EstimMethod]{fmmr6::EstimMethod}} -> \code{mle}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fit}{\code{mle$fit()}}
\item \href{#method-clone}{\code{mle$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="EstimMethod" data-id="gen_constraint">}\href{../../fmmr6/html/EstimMethod.html#method-gen_constraint}{\code{fmmr6::EstimMethod$gen_constraint()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="EstimMethod" data-id="gen_start">}\href{../../fmmr6/html/EstimMethod.html#method-gen_start}{\code{fmmr6::EstimMethod$gen_start()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="EstimMethod" data-id="initialize">}\href{../../fmmr6/html/EstimMethod.html#method-initialize}{\code{fmmr6::EstimMethod$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fmmr6" data-topic="EstimMethod" data-id="partial">}\href{../../fmmr6/html/EstimMethod.html#method-partial}{\code{fmmr6::EstimMethod$partial()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the MLE.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mle$fit(algo = "mle", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{algo}}{(\code{character(1)}) \cr
The algorithm used. The default is \code{mle}.}

\item{\code{...}}{(\code{list()}) \cr
Other related parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
